Renames FormOptions and tweaks AmendProtocol layout
